Work towards major changes and slight improvements to performance and usability
Chunk meshing optimizations.
Moved to extension methods for converting Vector3 types.
Cleanup here and there.
Material system refactor.
Work on rendering optimizations

+++ Engine/MathLib/MathHelper.cs
@@ -4,9 +4,9 @@ namespace Engine.MathLib
 {
     public static class MathHelper
     {
-        public static double DegreesToRadians(double yRotation)
+        public static float DegreesToRadians(double yRotation)
         {
-            return yRotation * (Math.PI / 180);
+            return (float) (yRotation * (Math.PI / 180));
         }
         
         public static float DegreesToRadians(float yRotation)
